project/requirements.txt
Django==3.0.7
psycopg2-binary==2.8.5
djangorestframework==3.11.0
django-import-export==2.3.0